Added some "Commentary".
Minor changes in `recentf-open-more-files' and `recentf-edit-list' to
move the point at the top of the file list. This behaviour is
consistent with the menu one when the list contains a lot of files.

`recentf-cleanup' now displays the number of items removed from the
list.

New `recentf-relative-filter' to show filenames relative to
`default-directory'.