(swbuff-window-lines): Use `swbuff-count-lines' instead of `count-lines'.
(swbuff-ignore): Alias of `ignore' used to prevent discarding the
status window on some mouse event.

In compatibility code: Use mouse-1, mouse-3 on mode line buffer
identification to respectively switch to previous or next buffer.  And
mouse-2 to kill the current buffer.

(swbuff-timer): New variable.  Timer used to discard the status
window.

(swbuff-show-status-window, swbuff-pre-command-hook):  Use timer to
discard the status window.

(swbuff-pre-command-hook):  Added `swbuff-kill-this-buffer' and
`swbuff-ignore' to the list of commands that do not discard the status
window.

(swbuff-kill-this-buffer): New command to kill the current buffer.